requirements.txt
Django==1.9
django-rest-swagger==0.3.5
djangorestframework==3.3.3
drfdocs==0.0.9
httpie==0.9.3
Markdown==2.6.6
Pygments==2.1.3
PyYAML==3.11
requests==2.9.1
wheel==0.24.0
